azure-function-match/requirements.txt
azure-functions
fastapi
pydantic
scikit-learn
sentence-transformers
